Add interactive execution of all commands, and on specific services
Change `docker-compose-execute-command` to accept both a command and a
service, and prompt for them when called interactively.

The commands use a predefined list (gotten from copy-pasting from
`docker-compose help`, hehe) and the services use `docker-compose
config --services` to get a list of the services.

A special string `docker-compose-all-services-token`, which defaults
to `"*all services*"` is prepended to the services list. When
selected, it signals that the command is to be run on all
services (e.g. with no further arguments to the command line).